# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RegisterDto" } }
          }
        },
        "responses": {
          "201": { "description": "User successfully registered." },
          "409": { "description": "Email already exists." }
        },
        "summary": "Register a new user",
        "tags": ["auth"]
      }
    },
    "/api/auth/verify": {
      "get": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [
          { "name": "token", "required": true, "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "summary": "Verify email address",
        "tags": ["auth"]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginDto" } }
          }
        },
        "responses": {
          "200": { "description": "Login successful." },
          "401": { "description": "Invalid credentials." }
        },
        "summary": "Login user",
        "tags": ["auth"]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Logout user",
        "tags": ["auth"]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshTokens",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RefreshTokenDto" } }
          }
        },
        "responses": { "200": { "description": "" } },
        "summary": "Refresh access token",
        "tags": ["auth"]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ChangePasswordDto" } }
          }
        },
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Change password",
        "tags": ["auth"]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ForgotPasswordDto" } }
          }
        },
        "responses": { "200": { "description": "" } },
        "summary": "Request password reset",
        "tags": ["auth"]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ResetPasswordDto" } }
          }
        },
        "responses": { "200": { "description": "" } },
        "summary": "Reset password using token",
        "tags": ["auth"]
      }
    },
...
    "/api/admin/users": {
      "get": {
        "operationId": "AdminUsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "minimum": 1, "default": 1, "type": "number" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": { "minimum": 1, "default": 10, "type": "number" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Get all users",
        "tags": ["admin/users"]
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "operationId": "AdminUsersController_findOne",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get user detail",
        "tags": ["admin/users"]
      }
    },
    "/api/admin/users/{id}/role": {
      "patch": {
        "operationId": "AdminUsersController_updateRole",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateUserRoleDto" } }
          }
        },
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Update user role",
        "tags": ["admin/users"]
      }
    },
...
    "/api/admin/plans": {
      "get": {
        "operationId": "AdminPlanController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Plan" } }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get all subscription plans",
        "tags": ["admin/plans"]
      },
      "post": {
        "operationId": "AdminPlanController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Plan" } } }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create new plan",
        "tags": ["admin/plans"]
      }
    },
    "/api/admin/plans/{id}": {
      "get": {
        "operationId": "AdminPlanController_findOne",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Plan" } } }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get plan detail",
        "tags": ["admin/plans"]
      },
      "patch": {
        "operationId": "AdminPlanController_update",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Plan" } } }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update plan (e.g., changes limits)",
        "tags": ["admin/plans"]
      }
    },
    "/api/admin/prompts": {
      "post": {
        "operationId": "AdminPromptController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Prompt" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create a new prompt",
        "tags": ["Admin Prompts"]
      },
      "get": {
        "operationId": "AdminPromptController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Prompt" } }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "List all prompts",
        "tags": ["Admin Prompts"]
      }
    },
    "/api/admin/prompts/{id}": {
      "patch": {
        "operationId": "AdminPromptController_update",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Prompt" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update a prompt",
        "tags": ["Admin Prompts"]
      },
      "delete": {
        "operationId": "AdminPromptController_remove",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Delete a prompt",
        "tags": ["Admin Prompts"]
      }
    },
...
  },
  "info": {
    "title": "AI Job Application Context API",
    "description": "The AI Job Application API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "minLength": 6, "example": "password123" }
        },
        "required": ["email", "password"]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "example": "password123" }
        },
        "required": ["email", "password"]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": { "refreshToken": { "type": "string" } },
        "required": ["refreshToken"]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": { "type": "string" },
          "newPassword": { "type": "string", "minLength": 6 }
        },
        "required": ["oldPassword", "newPassword"]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" }
        },
        "required": ["email"]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": { "type": "string", "minLength": 6, "example": "newPassword123" },
          "token": { "type": "string" }
        },
        "required": ["newPassword", "token"]
      },
...
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": { "role": { "enum": ["user", "admin"], "type": "string" } },
        "required": ["role"]
      },
      "Plan": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "slug": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "currency": { "type": "string" },
          "interval": { "type": "object" },
          "limits": { "type": "object" },
          "isActive": { "type": "boolean" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "price",
          "currency",
          "interval",
          "limits",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "Prompt": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "key": { "type": "string" },
          "content": { "type": "string" },
          "description": { "type": "string" },
          "category": { "enum": ["CV", "JOB", "INTERVIEW", "OTHER"], "type": "string" },
          "isActive": { "type": "boolean" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "key",
          "content",
          "description",
          "category",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}